fix(category/productOffer): date validation, uri to json messages

+++ controllers/admin/productOfferController.js
@@ -1,7 +1,7 @@
 import Product from '../../models/productSchema.js';
 import ProductOffer from '../../models/productOfferSchema.js';
 import { statusCodes } from '../../helpers/statusCodes.js';
-const { BAD_REQUEST, OK, NOT_FOUND } = statusCodes;
+const { BAD_REQUEST, OK, NOT_FOUND, SERVER_ERROR } = statusCodes;
 
 const loadProductOffers = async (req, res) => {
     try {
@@ -43,11 +43,26 @@ const addProductOffer = async (req, res) => {
             return res.status(BAD_REQUEST).json({ success: false, message: "Fill all the fields"});
         }
 
+        const product = await Product.findById(productId);
+        if(!product) {
+            return res.status(BAD_REQUEST).json({ success: false, message: "Product not found"})
+        }
+
         const endDay = new Date(endDate);
-        endDay.setHours(23, 59, 59, 999)
+        endDay.setHours(23, 59, 59, 999);
 
         const startDay = new Date(startDate);
-        startDay.setHours(0, 0, 0, 0)
+
+        const today = new Date();
+        today.setHours(0, 0, 0, 0);
+
+        if(endDay && endDay < today || startDay && startDay < today) {
+            return res.status(BAD_REQUEST).json({ success: false, message: "Start date and End date should not be past"});
+        }
+
+        if(startDay &&  endDay && endDay < startDay) {
+            return res.status(BAD_REQUEST).json({ success: false, message: "End date should not be before start date"})
+        }
 
         await ProductOffer.deleteMany({ productId });
 
@@ -59,18 +74,19 @@ const addProductOffer = async (req, res) => {
             isActive: true
         });
 
-        return res.redirect("/admin/productOffers?success=Product offer added");
+        return res.status(200).json({ success: true, message: "Product offer added", redirect:'/admin/productOffers'});
     } catch (error) {
-        const err = new Error("Add product offer server error");
-        err.redirect = "/admin/productOffers?error=Server error";
-        throw err;
-    }
+        console.error(error);
+        return res.status(SERVER_ERROR).json({
+        success: false,
+        message: "Add product offer internal server error"
+        });
+  }
 }
 
 const editProductOffer = async (req, res) => {
     try {
         const { offerId, discountPercentage, startDate, endDate } = req.body;
-        console.log('Loading edit for offer:', offerId);
         
         if(!offerId){
             return res.status(BAD_REQUEST).json({ success: false, message: "Offer not found!"});
@@ -79,11 +95,21 @@ const editProductOffer = async (req, res) => {
             return res.status(BAD_REQUEST).json({ success: false, message: "Fill all the fields"});
         }
 
+        const today = new Date();
+        today.setHours(0, 0, 0, 0);
+
         const startDay = new Date(startDate);
-        startDay.setHours(0, 0, 0, 0);
 
         const endDay = new Date(endDate);
         endDay.setHours(23, 59, 59, 999);
+
+        if(startDay && startDay < today || endDay && endDay < today) {
+            return res.status(BAD_REQUEST).json({ success: false, message: "Start date and End date should not be past"})
+        }
+
+        if(startDay && endDay && endDay < startDay) {
+            return res.status(BAD_REQUEST).json({ success: false, message: "End date should not be before Start date"})
+        }
 
         await ProductOffer.findByIdAndUpdate(offerId, {
             discountPercentage,
@@ -91,11 +117,13 @@ const editProductOffer = async (req, res) => {
             endDate: endDay,
         });
 
-        return res.status(OK).json({ success: true, message: "Offer updated"});
+        return res.status(OK).json({ success: true, message: "Product Offer updated"});
     } catch (error) {
-        const err = new Error("Edit product offer server error");
-        err.redirect = "/admin/productOffers?error=Server error";
-        throw err;
+        console.error(error);
+        return res.status(SERVER_ERROR).json({
+        success: false,
+        message: "Edit product offer internal server error"
+        });
     }
 }
 
@@ -110,9 +138,11 @@ const activateProductOffer = async (req, res) => {
         
         return res.status(OK).json({ success: true, message: "Offer Activated"})
     } catch (error) {
-        const err = new Error("Activate product offer server error");
-        err.redirect = "/admin/productOffers?error=Server error";
-        throw err;
+        console.error(error);
+        return res.status(SERVER_ERROR).json({
+        success: false,
+        message: "Activate product offer internal server error"
+        });
     }
 }
 
@@ -123,12 +153,15 @@ const deactivateProductOffer = async (req, res) => {
             return res.status(BAD_REQUEST).json({ success: false, message: "Offer not found!"})
         }
         
-        await ProductOffer.findByIdAndUpdate(offerId, { isActive: false })
-        return res.status(OK).json({ success: true, message: "Offer Deactivated"})
+        await ProductOffer.findByIdAndUpdate(offerId, { isActive: false });
+
+        return res.status(OK).json({ success: true, message: "Offer Deactivated"});
     } catch (error) {
-        const err = new Error("Deactivate product offer server error");
-        err.redirect = "/admin/productOffers?error=Server error";
-        throw err;
+        console.error(error);
+        return res.status(SERVER_ERROR).json({
+        success: false,
+        message: "Deactivate product offer internal server error"
+        });
     }
 }
